Fixes after review comments
